+++ core/dbt/exceptions.py
@@ -162,6 +162,33 @@ class RPCKilledException(RuntimeException):
         }
 
 
+class RPCCompiling(RuntimeException):
+    CODE = 10010
+    MESSAGE = (
+        'RPC server is compiling the project, call the "status" method for'
+        ' compile status'
+    )
+
+
+class RPCLoadException(RuntimeException):
+    CODE = 10011
+    MESSAGE = (
+        'RPC server failed to compile project, call the "status" method for'
+        ' compile status'
+    )
+
+    def __init__(self, cause):
+        self.cause = cause
+        self.message = '{}: {}'.format(self.MESSAGE, self.cause['message'])
+        super().__init__(self.message)
+
+    def data(self):
+        return {
+            'cause': self.cause,
+            'message': self.message
+        }
+
+
 class DatabaseException(RuntimeException):
     CODE = 10003
     MESSAGE = "Database Error"
