+++ test/integration/048_rpc_test/data/expected_multi_source.csv
@@ -1,0 +1,4 @@
+id,first_name,color
+1,Larry,blue
+2,Curly,red
+3,Moe,green
