+++ test/integration/048_rpc_test/data/other_source_table.csv
@@ -1,0 +1,4 @@
+id,color
+1,blue
+2,red
+3,green
