+++ test/integration/048_rpc_test/data/other_table.csv
@@ -1,0 +1,4 @@
+id,first_name
+1,Larry
+2,Curly
+3,Moe
