+++ test/integration/048_rpc_test/error_models/model.sql
@@ -1,0 +1,1 @@
+select * from {{ source('test_source', 'test_table') }}
