+++ test/integration/048_rpc_test/error_models/schema.yml
@@ -1,0 +1,12 @@
+version: 2
+sources:
+  - name: test_source
+    loader: custom
+    freshness:
+      warn_after: {count: 10, period: hour}
+      error_after: {count: 1, period: day}
+    schema: invalid
+    tables:
+      - name: test_table
+        identifier: source
+        loaded_at_field: updated_at
