+++ test/integration/048_rpc_test/malformed_models/descendant_model.sql
@@ -1,0 +1,1 @@
+select * from {{ source('test_source', 'test_table') }}
