+++ test/integration/048_rpc_test/models/ephemeral_model.sql
@@ -1,0 +1,3 @@
+{{ config(materialized='ephemeral') }}
+
+select 1 as id
