+++ test/integration/048_rpc_test/models/multi_source_model.sql
@@ -1,0 +1,2 @@
+select * from {{ source('test_source', 'other_test_table')}}
+	join {{ source('other_source', 'test_table')}} using (id)
