+++ test/integration/048_rpc_test/models/nonsource_descendant.sql
@@ -1,0 +1,1 @@
+select * from {{ schema }}.source
